+++ examples/master_querant.rs
@@ -1,6 +1,6 @@
 
 use std::env;
-use gamedig::{aliens, asrd, csgo, css, dods, gm, hl2dm, ins, insmic, inss, l4d, l4d2, tf2, ts};
+use gamedig::{aliens, asrd, cscz, csgo, css, dods, gm, hl2dm, ins, insmic, inss, l4d, l4d2, tf2, ts};
 use gamedig::protocols::valve;
 use gamedig::protocols::valve::App;
 
@@ -39,7 +39,9 @@ fn main() {
         "l4d" => println!("{:?}", l4d::query(ip, port)),
         "l4d2" => println!("{:?}", l4d2::query(ip, port)),
         "ts" => println!("{:?}", ts::query(ip, port)),
-        "_" => println!("{:?}", valve::query(ip, 27015, App::Source(None), None)),
+        "cscz" => println!("{:?}", cscz::query(ip, port)),
+        "_src" => println!("{:?}", valve::query(ip, 27015, App::Source(None), None)),
+        "_gld" => println!("{:?}", valve::query(ip, 27015, App::GoldSrc, None)),
         _ => panic!("Undefined game: {}", args[1])
     };
 }